+++ atomica/model.py
@@ -226,7 +226,7 @@ class Variable:
 
 
 class Compartment(Variable):
-    """ A class to wrap up data for one compartment within a cascade network. """
+    """A class to wrap up data for one compartment within a cascade network."""
 
     def __init__(self, pop, name):
         Variable.__init__(self, pop=pop, id=(pop.name, name))
@@ -957,7 +957,7 @@ class TimedCompartment(Compartment):
 
 
 class Characteristic(Variable):
-    """ A characteristic represents a grouping of compartments. """
+    """A characteristic represents a grouping of compartments."""
 
     def __init__(self, pop, name):
         # includes is a list of Compartments, whose values are summed
@@ -1573,6 +1573,25 @@ class Population:
     def __repr__(self):
         return '%s "%s"' % (self.__class__.__name__, self.name)
 
+    def __contains__(self, item: str) -> bool:
+        """
+        Check whether variables can be resolved
+
+        This function returns True if ``Population.get_variable(item)`` would result
+        in some variables being returned. This facilitates checking whether a population
+        contains particular variables - for example, transfer links may only exist in
+        a subset of the model populations, or if population types are being used, not all
+        variables will be defined in any one population.
+
+        :param item: The name of a parameter or a link specification (supported by ``get_variable``)
+        :return: True if ``Population.get_variable`` would return at least one variable, otherwise False
+        """
+        try:
+            self.get_variable(item)
+            return True
+        except NotFoundError:
+            return False
+
     def unlink(self):
         if not self.is_linked:
             return
@@ -1654,8 +1673,7 @@ class Population:
             return [self.par_lookup[name]]
         elif name in self.link_lookup:
             return self.link_lookup[name]
-        elif ":" in name:
-
+        elif ":" in name and not name.endswith(":flow"):
             name_tokens = name.split(":")
             if len(name_tokens) == 2:
                 name_tokens.append("")
@@ -1678,14 +1696,14 @@ class Population:
             raise NotFoundError(f"Object '{name}' not found in population '{self.name}'")
 
     def get_comp(self, comp_name):
-        """ Allow compartments to be retrieved by name rather than index. Returns a Compartment. """
+        """Allow compartments to be retrieved by name rather than index. Returns a Compartment."""
         try:
             return self.comp_lookup[comp_name]
         except KeyError:
             raise NotFoundError(f"Compartment {comp_name} not found")
 
     def get_links(self, name) -> list:
-        """ Retrieve Links. """
+        """Retrieve Links."""
         # Links can be looked up by parameter name or by link name, unlike get_variable. This is because
         # get_links() is guaranteed to return a list of Link objects
         # As opposed to get_variable which would retrieve the Parameter for 'doth rate' and the Links for 'z'
@@ -1697,14 +1715,14 @@ class Population:
             raise NotFoundError("Object '{0}' not found.".format(name))
 
     def get_charac(self, charac_name):
-        """ Allow dependencies to be retrieved by name rather than index. Returns a Variable. """
+        """Allow dependencies to be retrieved by name rather than index. Returns a Variable."""
         try:
             return self.charac_lookup[charac_name]
         except KeyError:
             raise NotFoundError(f"Characteristic {charac_name} not found")
 
     def get_par(self, par_name):
-        """ Allow dependencies to be retrieved by name rather than index. Returns a Variable. """
+        """Allow dependencies to be retrieved by name rather than index. Returns a Variable."""
         try:
             return self.par_lookup[par_name]
         except KeyError:
@@ -1941,7 +1959,7 @@ class Population:
 
 
 class Model:
-    """ A class to wrap up multiple populations within model and handle cross-population transitions. """
+    """A class to wrap up multiple populations within model and handle cross-population transitions."""
 
     def __init__(self, settings, framework, parset, progset=None, program_instructions=None):
 
@@ -2101,12 +2119,12 @@ class Model:
         return new
 
     def get_pop(self, pop_name):
-        """ Allow model populations to be retrieved by name rather than index. """
+        """Allow model populations to be retrieved by name rather than index."""
         pop_index = self._pop_ids[pop_name]
         return self.pops[pop_index]
 
     def build(self, parset):
-        """ Build the full model. """
+        """Build the full model."""
 
         # First construct populations
         for k, (pop_name, pop_label, pop_type) in enumerate(zip(parset.pop_names, parset.pop_labels, parset.pop_types)):
@@ -2338,7 +2356,7 @@ class Model:
         self._exec_order = exec_order
 
     def process(self) -> None:
-        """ Run the full model. """
+        """Run the full model."""
 
         assert self._t_index == 0  # Only makes sense to process a simulation once, starting at ti=0 - this might be relaxed later on
         self._set_exec_order()  # Set the execution order again in case the user has updated the parameters etc. It is critically important that this is correct during integration